add an assert to extending buffer and change argument type

+++ src/aostr.h
@@ -22,7 +22,7 @@ aoStr *aoStrAlloc(size_t capacity);
 aoStr *aoStrNew(void);
 void aoStrRelease(aoStr *buf);
 
-int aoStrExtendBuffer(aoStr *buf, unsigned int additional);
+int aoStrExtendBuffer(aoStr *buf, size_t additional);
 void aoStrToLowerCase(aoStr *buf);
 void aoStrPutChar(aoStr *buf, char ch);
 void aoStrRepeatChar(aoStr *buf, char ch, int times);
